requirements.txt
albucore==0.0.23 ; python_version >= "3.10"
albumentations==2.0.5 ; python_version >= "3.10"
annotated-types==0.7.0 ; python_version >= "3.10"
asttokens==3.0.0 ; python_version >= "3.10"
colorama==0.4.6 ; python_version >= "3.10"
contourpy==1.3.1 ; python_version >= "3.10"
cycler==0.12.1 ; python_version >= "3.10"
executing==2.2.0 ; python_version >= "3.10"
filelock==3.17.0 ; python_version >= "3.10"
fonttools==4.56.0 ; python_version >= "3.10"
fsspec==2025.2.0 ; python_version >= "3.10"
icecream==2.1.4 ; python_version >= "3.10"
jinja2==3.1.6 ; python_version >= "3.10"
joblib==1.4.2 ; python_version >= "3.10"
kiwisolver==1.4.8 ; python_version >= "3.10"
markupsafe==3.0.2 ; python_version >= "3.10"
matplotlib==3.10.1 ; python_version >= "3.10"
mpmath==1.3.0 ; python_version >= "3.10"
networkx==3.4.2 ; python_version >= "3.10"
numpy==2.2.3 ; python_version >= "3.10"
nvidia-cublas-cu12==12.4.5.8 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-cuda-cupti-cu12==12.4.127 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-cuda-nvrtc-cu12==12.4.127 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-cuda-runtime-cu12==12.4.127 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-cudnn-cu12==9.1.0.70 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-cufft-cu12==11.2.1.3 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-curand-cu12==10.3.5.147 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-cusolver-cu12==11.6.1.9 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-cusparse-cu12==12.3.1.170 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-cusparselt-cu12==0.6.2 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-nccl-cu12==2.21.5 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-nvjitlink-cu12==12.4.127 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
nvidia-nvtx-cu12==12.4.127 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
opencv-python-headless==4.11.0.86 ; python_version >= "3.10"
packaging==24.2 ; python_version >= "3.10"
pandas==2.2.3 ; python_version >= "3.10"
pillow==11.1.0 ; python_version >= "3.10"
pydantic-core==2.27.2 ; python_version >= "3.10"
pydantic==2.10.6 ; python_version >= "3.10"
pygments==2.19.1 ; python_version >= "3.10"
pyparsing==3.2.1 ; python_version >= "3.10"
python-dateutil==2.9.0.post0 ; python_version >= "3.10"
pytz==2025.1 ; python_version >= "3.10"
pyyaml==6.0.2 ; python_version >= "3.10"
scikit-learn==1.6.1 ; python_version >= "3.10"
scipy==1.15.2 ; python_version >= "3.10"
setuptools==75.8.2 ; python_version >= "3.12"
simsimd==6.2.1 ; python_version >= "3.10"
six==1.17.0 ; python_version >= "3.10"
stringzilla==3.12.2 ; python_version >= "3.10"
sympy==1.13.1 ; python_version >= "3.10"
threadpoolctl==3.5.0 ; python_version >= "3.10"
torch==2.6.0 ; python_version >= "3.10"
torchvision==0.21.0 ; python_version >= "3.10"
tqdm==4.67.1 ; python_version >= "3.10"
triton==3.2.0 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.10"
typing-extensions==4.12.2 ; python_version >= "3.10"
tzdata==2025.1 ; python_version >= "3.10"
